+++ product-service/src/main/java/com/example/product/dto/admin/OptionItemDto.java
@@ -3,11 +3,13 @@ package com.example.product.dto.admin;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Builder
 public class OptionItemDto {
 
     private Integer optionId;
     private String optionName;
-    private Integer defaultPrice;
+    private BigDecimal defaultPrice;
 }
