+++ product-service/src/main/java/com/example/product/repository/MenuOptionRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface MenuOptionRepository extends JpaRepository<MenuOption, Long> {
-    List<MenuOption> findByMenu_MenuCode(Integer menuCode);
-    void deleteByMenu_MenuCode(Integer menuCode);
+    List<MenuOption> findByMenu_MenuCode(Long menuCode);
+    void deleteByMenu_MenuCode(Long menuCode);
 }
