+++ product-service/src/main/java/com/example/product/repository/NutritionRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface NutritionRepository extends JpaRepository<Nutrition, Integer> {
-    Optional<Nutrition> findByMenu_MenuCode(Integer menuCode);
-    void deleteByMenu_MenuCode(Integer menuCode);
+public interface NutritionRepository extends JpaRepository<Nutrition, Long> {
+    Optional<Nutrition> findByMenu_MenuCode(Long menuCode);
+    void deleteByMenu_MenuCode(Long menuCode);
 }
