+++ product-service/src/main/java/com/example/product/repository/RecipeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface RecipeRepository extends JpaRepository<Recipe, Integer> {
-    List<Recipe> findByMenu_MenuCode(Integer menuCode);
-    void deleteByMenu_MenuCode(Integer menuCode);
+    List<Recipe> findByMenu_MenuCode(Long menuCode);
+    void deleteByMenu_MenuCode(Long menuCode);
 }
